feat: 优化参数

+++ coupon/entity/entity.go
@@ -1,8 +1,8 @@
 package entity
 
 type CouponUseReq struct {
-	UserID         int64   `json:"user_id"`
-	OrderID        int64   `json:"order_id"`
-	CouponID       int64   `json:"coupon_id"`
-	DiscountAmount float64 `json:"discount_amount"`
+	UserID         int64 `json:"user_id"`
+	OrderID        int64 `json:"order_id"`
+	CouponID       int64 `json:"coupon_id"`
+	DiscountAmount int64 `json:"discount_amount"`
 }
